Health API has moved to separate module
