+++ client-apis/health-api/src/main/java/org/wildfly/swarm/health/HealthStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.monitor;
+package org.wildfly.swarm.health;
 
 import java.util.HashMap;
 import java.util.Map;
