+++ client-apis/health-api/src/main/java/org/wildfly/swarm/health/Status.java
@@ -13,18 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.monitor;
+package org.wildfly.swarm.health;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.util.Optional;
 
 /**
- * Indicates that a JAX-RS method should be used check liveliness
+ * @author Heiko Braun
+ * @since 23/03/16
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Health {
+public interface Status {
+
+    String toJson();
+
+    enum State { UP, DOWN }
+
+    State getState();
+
+    Optional<String> getMessage();
 
 }
