--- fractions/monitor/src/main/java/org/wildfly/swarm/monitor/Status.java
@@ -1,34 +1,0 @@
-/**
- * Copyright 2015-2016 Red Hat, Inc, and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wildfly.swarm.monitor;
-
-import java.util.Optional;
-
-/**
- * @author Heiko Braun
- * @since 23/03/16
- */
-public interface Status {
-
-    String toJson();
-
-    enum State { UP, DOWN }
-
-    State getState();
-
-    Optional<String> getMessage();
-
-}
