+++ fractions/monitor/src/main/java/org/wildfly/swarm/monitor/runtime/InstallMonitorFilter.java
@@ -18,14 +18,9 @@ package org.wildfly.swarm.monitor.runtime;
 import javax.enterprise.context.ApplicationScoped;
 
 import org.jboss.shrinkwrap.api.Archive;
-import org.jboss.shrinkwrap.api.ArchivePath;
-import org.jboss.shrinkwrap.api.ArchivePaths;
-import org.jboss.shrinkwrap.api.asset.Asset;
-import org.jboss.shrinkwrap.api.asset.ClassAsset;
-import org.jboss.shrinkwrap.impl.base.asset.AssetUtil;
-import org.jboss.shrinkwrap.impl.base.path.BasicPath;
 import org.wildfly.swarm.spi.api.ArchivePreparer;
-import org.wildfly.swarm.spi.api.JARArchive;
+import org.wildfly.swarm.spi.api.internal.SwarmInternalProperties;
+import org.wildfly.swarm.undertow.WARArchive;
 
 /**
  * @author Ken Finnigan
@@ -33,24 +28,14 @@ import org.wildfly.swarm.spi.api.JARArchive;
 @ApplicationScoped
 public class InstallMonitorFilter implements ArchivePreparer {
 
-    //private static String HEALTH_RESPONSE_FILTER_CLASS_NAME = "org.wildfly.swarm.monitor.runtime.HealthResponseFilter";
-
-    /**
-     * Path to the WEB-INF inside of the Archive.
-     */
-    private static final ArchivePath PATH_WEB_INF = ArchivePaths.create("WEB-INF");
-
-    /**
-     * Path to the classes inside of the Archive.
-     */
-    private static final ArchivePath PATH_CLASSES = ArchivePaths.create(PATH_WEB_INF, "classes");
-
     @Override
     public void prepareArchive(Archive<?> archive) {
-        JARArchive jarArchive = archive.as(JARArchive.class);
-
-        Asset resource = new ClassAsset(HealthResponseFilter.class);
-        ArchivePath location = new BasicPath(PATH_CLASSES, AssetUtil.getFullPathForClassResource(HealthResponseFilter.class));
-        jarArchive.add(resource, location);
+        try {
+            WARArchive warArchive = archive.as(WARArchive.class);
+            warArchive.addDependency("org.wildfly.swarm:health-api:jar:" + System.getProperty(SwarmInternalProperties.VERSION));
+            warArchive.findWebXmlAsset().setContextParam("resteasy.scan", "true");
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to install health processor", e);
+        }
     }
 }
