+++ testsuite/testsuite-jaxrs/src/main/java/org/wildfly/swarm/jaxrs/SuccessfulChecks.java
@@ -18,8 +18,8 @@ package org.wildfly.swarm.jaxrs;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 
-import org.wildfly.swarm.monitor.Health;
-import org.wildfly.swarm.monitor.HealthStatus;
+import org.wildfly.swarm.health.Health;
+import org.wildfly.swarm.health.HealthStatus;
 
 /**
  * @author Heiko Braun
